add structureCopy
